// src/common/enums/ticket.enum.ts
export enum UserRole {
    MANAGER = 'MANAGER',
    SUPPORT = 'SUPPORT',
    USER = 'USER',
}

export enum TicketStatus {
    OPEN = 'OPEN',
    IN_PROGRESS = 'IN_PROGRESS',
    RESOLVED = 'RESOLVED',
    CLOSED = 'CLOSED',
}

export enum TicketPriority {
    LOW = 'LOW',
    MEDIUM = 'MEDIUM',
    HIGH = 'HIGH',
}